move grid generation out of grid search file
