+++ fairlearn/reductions/_grid_search/_grid_generation.py
@@ -1,0 +1,65 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+import numpy as np
+import pandas as pd
+
+
+class _GridGenerator:
+    """A generator of a grid of points with a bounded L1 norm."""
+
+    def __init__(self, grid_size, grid_limit, pos_basis, neg_basis, neg_allowed, force_L1_norm):
+        # grid parameters
+        self.dim = len(pos_basis.columns)
+        self.neg_allowed = neg_allowed
+        self.force_L1_norm = force_L1_norm
+
+        # true dimensionality of the grid
+        if self.force_L1_norm:
+            true_dim = self.dim - 1
+        else:
+            true_dim = self.dim
+
+        # a conservative lower bound on the scaling parameter of the grid
+        n_units = (float(grid_size) / (2.0**neg_allowed.sum())) ** (1.0 / true_dim) - 1   # noqa: E501
+        n_units = int(np.floor(n_units))
+        if n_units < 0:
+            n_units = 0
+
+        # find the grid of size at least "size" and save the first "size" entries
+        while True:
+            int_grid = self.build_integer_grid(n_units)
+            if len(int_grid) >= grid_size:
+                # re-scale the integer grid, separate into positive and negative parts
+                pos_coefs = pd.DataFrame(self.accumulator[:grid_size]).T * (float(grid_limit) / n_units)   # noqa: E501
+                neg_coefs = -pos_coefs.copy()
+                pos_coefs[pos_coefs < 0] = 0.0
+                neg_coefs[neg_coefs < 0] = 0.0
+                # convert the grid of basis coefficients into a grid of lambda vectors
+                self.grid = pos_basis.dot(pos_coefs) + neg_basis.dot(neg_coefs)
+                break
+            n_units = n_units + 1
+
+    def build_integer_grid(self, n_units):
+        # initialize working variables for the grid accumulation
+        self.entry = np.zeros(self.dim)
+        self.accumulator = []
+        # recursively create the integer grid
+        self.accumulate_integer_grid(0, n_units)
+        return self.accumulator
+
+    def accumulate_integer_grid(self, index, max_val):
+        if index == self.dim:
+            self.accumulator.append(self.entry.copy())
+        else:
+            if (index == self.dim - 1) and (self.force_L1_norm):
+                if self.neg_allowed[index] and max_val > 0:
+                    values = [-max_val, max_val]
+                else:
+                    values = [max_val]
+            else:
+                min_val = -max_val if self.neg_allowed[index] else 0
+                values = range(min_val, max_val + 1)
+            for current_value in values:
+                self.entry[index] = current_value
+                self.accumulate_integer_grid(index + 1, max_val - abs(current_value))   # noqa: E501
